refactor: Split manifest crate
