+++ crates/iceberg/src/expr/visitors/expression_evaluator.rs
@@ -19,7 +19,8 @@ use fnv::FnvHashSet;
 
 use super::bound_predicate_visitor::{visit, BoundPredicateVisitor};
 use crate::expr::{BoundPredicate, BoundReference};
-use crate::spec::{DataFile, Datum, PrimitiveLiteral, Struct};
+use crate::spec::data_file::DataFile;
+use crate::spec::{Datum, PrimitiveLiteral, Struct};
 use crate::{Error, ErrorKind, Result};
 
 /// Evaluates a [`DataFile`]'s partition [`Struct`] to check
@@ -257,10 +258,10 @@ mod tests {
         predicate, BinaryExpression, Bind, BoundPredicate, Predicate, PredicateOperator, Reference,
         UnaryExpression,
     };
+    use crate::spec::data_file::{DataContentType, DataFile, DataFileFormat};
     use crate::spec::{
-        DataContentType, DataFile, DataFileFormat, Datum, Literal, NestedField, PartitionSpec,
-        PartitionSpecRef, PrimitiveType, Schema, SchemaRef, Struct, Transform, Type,
-        UnboundPartitionField,
+        Datum, Literal, NestedField, PartitionSpec, PartitionSpecRef, PrimitiveType, Schema,
+        SchemaRef, Struct, Transform, Type, UnboundPartitionField,
     };
     use crate::Result;
