+++ crates/iceberg/src/writer/base_writer/equality_delete_writer.rs
@@ -26,7 +26,8 @@ use parquet::arrow::PARQUET_FIELD_ID_META_KEY;
 
 use crate::arrow::record_batch_projector::RecordBatchProjector;
 use crate::arrow::schema_to_arrow_schema;
-use crate::spec::{DataFile, SchemaRef, Struct};
+use crate::spec::data_file::{DataContentType, DataFile};
+use crate::spec::{SchemaRef, Struct};
 use crate::writer::file_writer::{FileWriter, FileWriterBuilder};
 use crate::writer::{IcebergWriter, IcebergWriterBuilder};
 use crate::{Error, ErrorKind, Result};
@@ -150,7 +151,7 @@ impl<B: FileWriterBuilder> IcebergWriter for EqualityDeleteFileWriter<B> {
                 .await?
                 .into_iter()
                 .map(|mut res| {
-                    res.content(crate::spec::DataContentType::EqualityDeletes);
+                    res.content(DataContentType::EqualityDeletes);
                     res.equality_ids(self.equality_ids.iter().copied().collect_vec());
                     res.partition(self.partition_value.clone());
                     res.build().expect("msg")
@@ -184,10 +185,8 @@ mod test {
 
     use crate::arrow::{arrow_schema_to_schema, schema_to_arrow_schema};
     use crate::io::{FileIO, FileIOBuilder};
-    use crate::spec::{
-        DataFile, DataFileFormat, ListType, MapType, NestedField, PrimitiveType, Schema,
-        StructType, Type,
-    };
+    use crate::spec::data_file::{DataFile, DataFileFormat};
+    use crate::spec::{ListType, MapType, NestedField, PrimitiveType, Schema, StructType, Type};
     use crate::writer::base_writer::equality_delete_writer::{
         EqualityDeleteFileWriterBuilder, EqualityDeleteWriterConfig,
     };
